pyproject.toml
[project]
name = "redis-advanced-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "flower>=2.0.1",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
]
